Revert "making interaction betwwen statusbar and normalbar"
